--config-file should always have priority over the CLI arguments
Otherwise it's all very confusing. Also warn the user if something
weird is happening (like passing options both on the CLI and in the
config file).

All of this was very clever, but super confusing :-)

Also: Cmdliner does not distinguish between default argument and
the lack of arguments. So the tricks we were playing won't work
if we remove our custom partial implementation of Cmdliner to use
the upstream version.